feature(common-service) : device, device-category 분리
