+++ packages/common-services/src/services/device.ts
@@ -3,14 +3,12 @@ import { api } from '@plug/api-hooks/core';
 import type {
   GsDeviceResponse,
   GsDeviceCreateRequest,
-  GsDeviceUpdateRequest,
-  DeviceCategoryResponse,
-  DeviceCategoryRequest,
-  DeviceCategoryAllResponse
+  GsDeviceUpdateRequest
 } from '../types/device';
 
 const DEVICE_END_POINT = `devices`;
-const DEVICE_CATEGORY_END_POINT = `device-categories`;
+// Re-export category-related APIs from the split file to preserve public API surface
+export * from './device-category';
 
 export const getDevices = async (): Promise<GsDeviceResponse[]> => {
   try {
@@ -49,51 +47,10 @@ export const useRemoveDeviceCategory = (deviceId: string) => {
   return useDelete(`${DEVICE_END_POINT}/${deviceId}/category`, { requireAuth: true });
 };
 
-export const useDeviceCategories = () => {
-  return useGet<DeviceCategoryAllResponse>(DEVICE_CATEGORY_END_POINT, { requireAuth: true });
-};
-
-export const useDeviceCategoryDetail = (categoryId: number) => {
-  return useGet<DeviceCategoryResponse>(`${DEVICE_CATEGORY_END_POINT}/${categoryId}`, { requireAuth: true });
-};
-
-export const useCreateDeviceCategory = () => {
-  return usePost<DeviceCategoryRequest>(DEVICE_CATEGORY_END_POINT, { requireAuth: true });
-};
-
-export const updateDeviceCategory = async (categoryId: number, data: DeviceCategoryRequest) => {  
-  return api.put(`${DEVICE_CATEGORY_END_POINT}/${categoryId}`, data, { requireAuth: true });
-};
-
-export const deleteDeviceCategory = async (categoryId: number) => {
-  return api.delete(`${DEVICE_CATEGORY_END_POINT}/${categoryId}`, { requireAuth: true });
-};
-
-export const useChildDeviceCategories = (categoryId: number) => {
-  return useGet<DeviceCategoryResponse[]>(`${DEVICE_CATEGORY_END_POINT}/${categoryId}/children`, { requireAuth: true });
-};
-
 export const useDevicesSWR = () => {
   return useSWRApi<GsDeviceResponse[]>(DEVICE_END_POINT, 'GET', { requireAuth: true });
 };
 
 export const useDeviceDetailSWR = (deviceId: string) => {
   return useSWRApi<GsDeviceResponse>(`${DEVICE_END_POINT}/${deviceId}`, 'GET', { requireAuth: true });
-};
-
-export const useDeviceCategoriesSWR = () => {
-  return useSWRApi<DeviceCategoryAllResponse>(DEVICE_CATEGORY_END_POINT, 'GET', { requireAuth: true });
-};
-
-export const useDeviceCategoryTree = () => {
-  const { data, error, isLoading, mutate } = useDeviceCategoriesSWR();
-  
-  return {
-    categories: data?.list || [],
-    maxDepth: data?.maxDepth || 0,
-    error,
-    isLoading,
-  mutate,
-  refresh: () => mutate(),
-  };
 };
