+++ packages/common-services/src/types/device.ts
@@ -1,23 +1,9 @@
-import { FileResponse } from "./file";
-export interface DeviceCategoryResponse {
-  id: number;
-  name: string;
-  parentId?: number;
-  children?: DeviceCategoryResponse[];
-  thumbnailFile?: FileResponse;
-  depth: number;
-}
-
-export interface DeviceCategoryRequest {
-  name: string;
-  parentId?: number | null;
-  thumbnailFileId?: number;
-}
-
-export interface DeviceCategoryAllResponse {
-  maxDepth: number;
-  list: DeviceCategoryResponse[];
-}
+import type { DeviceCategoryResponse } from './device-category';
+export type {
+  DeviceCategoryResponse,
+  DeviceCategoryRequest,
+  DeviceCategoryAllResponse,
+} from './device-category';
 
 export interface GsDeviceCreateRequest {
   id: string;
